feat: move get index data to dataset
