+++ cross-modal-search/app.py
@@ -4,9 +4,10 @@ __license__ = "Apache-2.0"
 import os
 
 import click
-
 from jina import Flow
-from jina import Document
+
+from dataset import input_index_data
+
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
 
@@ -28,34 +29,6 @@ def config(model_name):
         msg = f'Unsupported model {model_name}.'
         msg += 'Expected `clip` or `vse`.'
         raise ValueError(msg)
-
-
-def input_index_data(num_docs=None, batch_size=8, dataset_type='f30k'):
-    from dataset import get_data_loader
-    captions = 'dataset_flickr30k.json' if dataset_type == 'f30k' else 'captions.txt'
-    data_loader = get_data_loader(
-        root=os.path.join(cur_dir, f'data/{dataset_type}/images'),
-        captions=os.path.join(cur_dir, f'data/{dataset_type}/{captions}'),
-        split='test',
-        batch_size=batch_size,
-        dataset_type=dataset_type
-    )
-
-    for i, (images, captions) in enumerate(data_loader):
-        for image, caption in zip(images, captions):
-            with Document() as document_img:
-                document_img.buffer = image
-                document_img.modality = 'image'
-                document_img.mime_type = 'image/jpeg'
-
-            with Document() as document_caption:
-                document_caption.text = caption
-                document_caption.modality = 'text'
-                document_caption.mime_type = 'text/plain'
-            yield document_img, document_caption
-
-        if num_docs and (i + 1) * batch_size >= num_docs:
-            break
 
 
 @click.command()
